refactor(input): extract builders and render into separate modules
Move Input::new() and builder methods to builders.rs, View impl
to render.rs. No visibility changes needed (fields already pub(super)).
Tests remain in mod.rs (private field access).
